Accepte des grands entiers pour les chatId

+++ projects/models.py
@@ -137,7 +137,7 @@ class Task(models.Model):
 
 
 class Bot(models.Model):
-    chatId = models.IntegerField(default=0, null=True, blank=True)
+    chatId = models.BigIntegerField(default=0, null=True, blank=True)
     project = models.ForeignKey(Project, on_delete=models.CASCADE)
     verified = models.BooleanField(default=False)
     verifyToken = models.TextField(max_length=255, blank=True)
